requirements.txt
streamlit>=1.20.0
langchain>=0.0.267
langchain-google-genai>=0.0.5
faiss-cpu>=1.7.4
pypdf>=3.15.1
sentence-transformers>=2.2.2
openai>=1.3.0
google-generativeai>=0.3.0
huggingface-hub>=0.16.4
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
bs4

